fix(`/import`): ForceReply not cleared
This is not a perfect fix. Telegram does not allow clearing
ReplyKeyboardForceReply w/o sending a message w/ ReplyKeyboardHide, but
such a message cannot be edited, making anything in a dilemma. A
possible workaround is to delete the message itself, so that nothing can
be force-replied. That's exactly what the commit does.

Signed-off-by: Rongrong <i@rong.moe>