feat(scrips/aerich_helper): aerich helper for database migration
Creating migrations is never an easy task, especially when the ORM model
is deep inside Python packages and when the ORM configuration is
generated dynamically at runtime.

A helper script is created to solve the dilemma. It is also very dynamic
because all commands and parameters are dynamically inspected and
generated from aerich. Human-readable help messages (pass `-h`/`--help`
to see them) are also generated using the same methodology.

Signed-off-by: Rongrong <i@rong.moe>