fix(root): exit with 0 when disconnected by server side
bot.disconnected usually means the bot is logged out due to a network
error or Telegram DC degradation, so we should exit with a non-zero code
to indicate an error.

Now the bot will exit with exit-code 100 when disconnected. This aims to
avoid confusion when running the bot in a container or as a service.

Signed-off-by: Rongrong <i@rong.moe>