feat(monitor): improve style of periodic summary of monitoring subtasks
Improve the style of the periodic summary of monitoring subtasks to make
the relationship between different metrics clearer. In particular,
deferred and resubmitted subtasks are no longer counted as finished -
they will finish sooner or later, but not when they are counted.

The term "subtask" should have been used in some comments, logs, and
variable names, but in fact, "task" was used. This is also corrected.

Signed-off-by: Rongrong <i@rong.moe>