fix(`/sub`): monitoring not deferred as per server-side cache
Since b3837c07, monitoring tasks will be deferred when aggressive
server-side caches (e.g., Cloudflare and RSSHub, which make it futile to
check for updates before cache expiration) are detected. However, the
first monitoring task for a newly subscribed feed was not being
deferred. This has been fixed and the first monitoring task now waits
for the server-side cache to expire.

Signed-off-by: Rongrong <i@rong.moe>