fix(compat.utils): malformed `<` in `<code>` or `<pre>`
minify-html v0.16.0+ has changed its default behavior to be more
compliant with WHATWG specifications. The most significant change is the
addition of the option `allow_optimal_entities`, which makes it possible
to fix the issue by disabling it. Since it defaults to `False`, all we
need to do is not accept minified HTML from minify-html-onepass, since
the latter lacks the mentioned option. Other than that it's just random
refactoring.

Once minify-html-onepass adds the corresponding option or changes its
behavior, we can accept its minified HTML again.

Signed-off-by: Rongrong <i@rong.moe>